+++ lib/karafka/web/tracking/consumers/listeners/errors.rb
@@ -75,12 +75,6 @@ module Karafka
             # @param error [StandardError] error that occurred
             # @return [Array<String, String, String>] array with error name, message and backtrace
             def extract_error_info(error)
-              error_message = error.message.to_s.dup
-              error_message.force_encoding('utf-8')
-              error_message.scrub!
-
-              backtrace = (error.backtrace || [])
-
               app_root = "#{::Karafka.root}/"
 
               gem_home = if ENV.key?('GEM_HOME')
@@ -91,14 +85,26 @@ module Karafka
 
               gem_home = "#{gem_home}/"
 
+              backtrace = error.backtrace || []
               backtrace.map! { |line| line.gsub(app_root, '') }
               backtrace.map! { |line| line.gsub(gem_home, '') }
 
               [
                 error.class.name,
-                error_message,
+                extract_exception_message(error),
                 backtrace.join("\n")
               ]
+            end
+
+            # @param error [StandardError] error that occurred
+            # @return [String] formatted exception message
+            def extract_exception_message(error)
+              error_message = error.message.to_s[0, 10_000]
+              error_message.force_encoding('utf-8')
+              error_message.scrub! if error_message.respond_to?(:scrub!)
+              error_message
+            rescue StandardError
+              '!!! Error message extraction failed !!!'
             end
           end
         end